+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v2.5.0]
+
 ### Added
 - `google-analytics.js`, exporting functions to import Google Analytics script
 - `debounce()` function in `functions.js`